+++ inc/platform.hpp
@@ -27,7 +27,7 @@
 #ifndef GUID_DEFINED
 struct GUID
 {
-    unsigned long  Data1;
+    unsigned int   Data1;
     unsigned short Data2;
     unsigned short Data3;
     unsigned char  Data4[8];
